src/integration/agents/src/main/java/se/tink/backend/aggregation/agents/nxgen/it/openbanking/mediolanum/fetcher/MediolanumAccountFetcher.java
package se.tink.backend.aggregation.agents.nxgen.it.openbanking.mediolanum.fetcher;

import java.util.Collection;
import java.util.Optional;
import java.util.stream.Collectors;
import lombok.RequiredArgsConstructor;
import se.tink.backend.aggregation.agents.nxgen.it.openbanking.mediolanum.MediolanumApiClient;
import se.tink.backend.aggregation.nxgen.controllers.refresh.AccountFetcher;
import se.tink.backend.aggregation.nxgen.core.account.transactional.TransactionalAccount;

@RequiredArgsConstructor
public class MediolanumAccountFetcher implements AccountFetcher<TransactionalAccount> {

    private final MediolanumApiClient apiClient;
    private final AccountMapper accountMapper;

    @Override
    public Collection<TransactionalAccount> fetchAccounts() {
        return apiClient.fetchAccounts().getAccounts().stream()
                .map(accountMapper::toTinkAccount)
                .filter(Optional::isPresent)
                .map(Optional::get)
                .collect(Collectors.toList());
    }
}
